+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Household: reusable confirm popup, `getHouseholdMinimumAgeConfirmPopup`, to block the household section when no member reaches the new `ages.householdMinimumAge` config option (defaults to 16). Add it to the section's continue button: `confirmPopup: getHouseholdMinimumAgeConfirmPopup()` (fixes [#1858](https://github.com/chairemobilite/evolution/issues/1858)).
+- Journey: add the `_skipTripDiary` questionnaire time property, that, if set to `true` will skip the trip diary sections for the corresponding journey (fixes [#1855](https://github.com/chairemobilite/evolution/issues/1855)).
 
 ### Changed